backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "loguru>=0.7.2",
    "peewee>=3.17.6",
    "uvicorn>=0.31.0",
    "bcrypt>=4.0.1",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.1",
]
